snapshotter/requirements.txt
apache-airflow==1.9.0
black==18.9b0
flake8==3.7.6
web3==4.8.3